+++ libwdi/msapi_utf8.h
@@ -439,11 +439,11 @@ static __inline int ComboBox_GetLBTextU(HWND hCtrl, int index, char* lpString)
 	size = (int)SendMessageW(hCtrl, CB_GETLBTEXTLEN, (WPARAM)index, (LPARAM)0);
 	if (size < 0)
 		return size;
-	wlpString = (wchar_t*)calloc(size+1, sizeof(wchar_t));
+	wlpString = (wchar_t*)calloc((size_t)size + 1, sizeof(wchar_t));
 	size = (int)SendMessageW(hCtrl, CB_GETLBTEXT, (WPARAM)index, (LPARAM)wlpString);
 	err = GetLastError();
 	if (size > 0)
-		wchar_to_utf8_no_alloc(wlpString, lpString, size+1);
+		wchar_to_utf8_no_alloc(wlpString, lpString, size + 1);
 	wfree(lpString);
 	SetLastError(err);
 	return size;
